+++ TuPencaUy.CoreAPI/TuPencaUy.CoreAPI/Model/Requests/CreateEventRequest.cs
@@ -9,7 +9,7 @@ namespace TuPencaUy.Core.API.Model.Requests
     public required string Name { get; set; }
     public DateTime? StartDate { get; set; }
     public DateTime? EndDate { get; set; }
-    public float? Comission { get; set; }
+    public decimal? Comission { get; set; }
     public TeamTypeEnum TeamType { get; set; }
 
     public int Sport_id { get; set; }
