+++ TuPencaUy.CoreAPI/TuPencaUy.CoreAPI/Model/Requests/ModifyEventRequest.cs
@@ -8,7 +8,7 @@ namespace TuPencaUy.Core.API.Model.Requests
     public string? Name { get; set; }
     public DateTime? StartDate { get; set; }
     public DateTime? EndDate { get; set; }
-    public float? Comission { get; set; }
+    public decimal? Comission { get; set; }
     public TeamTypeEnum? TeamType { get; set; }
     public bool? Instantiable {  get; set; } 
   }
